+++ frontend/src/pages/Skill/CharacterSkillList.tsx
@@ -1,12 +1,13 @@
-import { Skill } from "../../types/skill";
-import { CharacterSkill } from "./CharacterSkill";
+import { CharacterSkill } from "../../types/skill";
+import { CharacterSkillItem } from "./CharacterSkillItem";
 import { Table, TableBody, TableCell, TableHead, TableRow } from "@mui/material";
 import { StyledTable } from "../../components/shared/Table.styled";
 import { useTranslation } from "react-i18next";
+import { height } from "@mui/system";
 
 interface IProps {
   charId: string;
-  skills: Record<string, Skill>;
+  skills: Record<string, CharacterSkill>;
   fetchCharHandler: () => void;
   canRemoveSkill: boolean;
 }
@@ -14,20 +15,29 @@ interface IProps {
 export const CharacterSkillList = ({ charId, skills, fetchCharHandler, canRemoveSkill }: IProps): JSX.Element => {
 
   const { t } = useTranslation("char");
-
   const charSkillItems = () => {
     const result: JSX.Element[] = [];
     if (skills) {
-      const skillMap: Map<string, Skill> = new Map(Object.entries(skills));
+      const skillMap: Map<string, CharacterSkill> = new Map(Object.entries(skills));
       skillMap.forEach((value, key) => {
         if (value != null) {
-          let skill = skillMap.get(key);
-          if (skill != null) {
-            result.push(<CharacterSkill key={skill.key} characterId={charId} skill={skill} fetchCharHandler={fetchCharHandler} canRemoveSkill={canRemoveSkill}/>);
+          let charSkill = skillMap.get(key);
+          if (charSkill != null) {
+            result.push(<CharacterSkillItem key={charSkill.skill.key} characterId={charId} charSkill={charSkill} fetchCharHandler={fetchCharHandler} canRemoveSkill={canRemoveSkill}/>);
           }
         }
       });
     }
+    let emptyRows = 3 - result.length < 1 ? 0 : 3 - result.length
+    for(let i = 0; i < emptyRows; i++) {
+      result.push(<TableRow key={i} style={{ height: "25px" }}>
+        <TableCell></TableCell>
+        <TableCell></TableCell>
+        <TableCell></TableCell>
+        <TableCell></TableCell>
+      </TableRow>);
+    }
+
     return result;
   };
 
@@ -37,10 +47,10 @@ export const CharacterSkillList = ({ charId, skills, fetchCharHandler, canRemove
       <Table>
         <TableHead>
           <TableRow>
-            {canRemoveSkill && <TableCell width={1}/>}
             <TableCell sx={{ fontWeight: 'bold' }}>{t("detail.skills.skills")}</TableCell>
             <TableCell sx={{ fontWeight: 'bold' }}>{t("detail.skills.value")}</TableCell>
             <TableCell sx={{ fontWeight: 'bold' }}>{t("detail.skills.exp")}</TableCell>
+            <TableCell width={1}/>
           </TableRow>
         </TableHead>
         <TableBody>